2052: Add tests
